ci(docs): Add CI job building/uploading/deploying documentation
If building documentation is successful, the generated HTML is uploaded
as a build artifact.  On a matching rev,  it is also deployed to GitHub
Pages.

We need to build  all protobuf files  to be able to  use sphinx autodoc
(otherwise `metricq` is not importable).   But we cannot run `setup.py`
to build protobuf files before having  all build dependencies installed
(e.g. `mypy-protobuf`).  These are specified in `pyproject.toml`.

Currently, no tool supports simply installing dependencies from this
file.  Options are:

1. `pip wheel ...`:
        Does not work; resolves and installs build dependencies from
        `pyproject.toml` in an isolated directory.  They are not
        accessible anymore after the wheel has been built.
2. `python -m pep517.build --source .`:
        Does not work for the same reasons as above.
        This would work if the source distribution this produces
        included the documentation source files (`./docs/*`).
3. Duplicate the list found in `pyproject.toml`:
        Rejected, DRY.
4. `pip install -r pyproject.toml`:
        Not supported by pip [1].

Option 4 leads us to this [2] *wonderful* solution:

        Parse `pyproject.toml` ourselves, tell `pip` to install the
        build dependencies we found.

Further advise on how to deal with Python packaging issues can be found
here [3].

[1] https://github.com/pypa/pip/issues/8049#issuecomment-633845028
[2] https://github.com/pypa/pip/issues/8049#issuecomment-633845028
[3] https://www.youtube.com/watch?v=vLaX8UvVUQw