Adds metricq-spy tool
This helper allows to get information about metrics. Just provide it 
with a list of metric regex and it will find any match using 
get_metrics. For every historic metric, it will also try to send a 
request in order to figure out the storage situation. Note: hic sunt 
dracones.