fix(types): Accept float instead of Real for durations
Raw duration values are annotated as `float` instead of the more
complicated `Real`.  From a type checker's perspective, `int` is
a subtype of `float`, so integers are implicitly valid parameters
in all of these functions.

The `Real` annotation is not what we want since it (ironically) does not
support multiplication and we need to scale the argument by some factor.