feat(tools): Add live data analysis tool
Inspect records of the given metric until the process is stopped with
SIGINT.  The tool then prints a histogram of the durations between two
consecutive data points and a histogram of the values of the data
points. In contrast to a tool similar to metricq-summary, the focus is
on one metric and it is blocking.