Fix version guessing for protoc
During the build process, protoc is invoked to build protobuf files.
These built files are deployed to users, and thus the users require a
compatible python-protobuf version.

For that, the required protobuf version is set during the build process
in setup.py. However, this version selection is not up to date with the
versioning scheme used by protobuf. protobuf keeps the major version
independent between languages, and minor/patch versions in sync.
E.g., as of the time of writing, python-protobuf 4.21.12 is compatible
with libprotoc 3.21.12.
see: https://protobuf.dev/news/2022-05-06/

However, it is not clear under which circumstances these versions are
compatible to each other. (I.e., even if minor/patch version are
different, the binary format may still be compatible.)
This makes correctly predicting the required (potentially unrealeased)
python-protobuf version impossible.

This patch takes a conservative approach: It manually ensures that
minor/patch versions align *during build*, and then requires the
*exact* python-protobuf used during building also during runtime.
As a consequence, only readily available python-protobuf versions should
be used during building, in order to prevent complications for users.