refactor: tweak isort
Only ignore parts of setup.py. This has two advantages: 1) If isort is applied to the file directly, e.g., by an eager IDE, it still doesn't sort the setuptools imports as needed. 2) The rest of the imports is sorted.

Also remove the configured ignore for docs/conf.py which seems to be fine.